Prevent loading of payment method models with wrong base type
Magento does not check that payment method object extend the base
abstract payment method class.  This means you could load a model
that does not have the expected methods and could casue a fatal error

This patch adds a simple check to ensure payment methods extend the abstract
class and if not adds a log message then skips the payment method.