Update payment method sort callback
With an updated check that ensures all method classes extend a base abstract,
we can use this to update the callback's method signature.

The sorting method does not need to return {-1,0,1} but any valid signed integer.
Taking advantage of this allows up to perform a quick subtraction and return the
result, dropping the 2 ternary and 2 gt/lt checks.